Separating the formatters into different files
